+++ wardenjs/__tests__/messages.test.js
@@ -1,3 +1,0 @@
-it('it works', async () => {
-  console.log('it works!');
-});