+++ wardenjs/src/codegen/gogoproto/bundle.ts
@@ -1,5 +1,0 @@
-//@ts-nocheck
-import * as _81 from "./gogo.js";
-export const gogoproto = {
-  ..._81
-};